fix: double->float

+++ include/color.h
@@ -6,7 +6,7 @@
 /*   By: stanaka2 <stanaka2@student.42tokyo.jp>     +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2026/06/19 01:09:47 by stanaka2          #+#    #+#             */
-/*   Updated: 2026/06/19 03:11:20 by stanaka2         ###   ########.fr       */
+/*   Updated: 2026/06/23 13:40:09 by stanaka2         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -15,9 +15,9 @@
 
 typedef struct s_color
 {
-	double	r;
-	double	g;
-	double	b;
+	float	r;
+	float	g;
+	float	b;
 }	t_color;
 
 enum e_color_channel
@@ -30,6 +30,6 @@ enum e_color_channel
 int		calc_rgb(t_color color);
 t_color	add_color(t_color a, t_color b);
 t_color	mul_color(t_color a, t_color b);
-t_color	scale_color(double t, t_color color);
+t_color	scale_color(float t, t_color color);
 
 #endif
